Move query files to a new folder
